+++ crates/iota-sdk-ffi/src/crypto/mod.rs
@@ -1,0 +1,5 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod passkey;
+pub mod simple;
