+++ crates/iota-sdk-ffi/src/crypto/simple.rs
