+++ crates/iota-sdk-ffi/src/types/crypto/passkey.rs
@@ -66,18 +66,3 @@ impl PasskeyAuthenticator {
         self.0.signature().into()
     }
 }
-
-#[derive(uniffi::Object)]
-pub struct PasskeyVerifier(iota_crypto::passkey::PasskeyVerifier);
-
-#[uniffi::export]
-impl PasskeyVerifier {
-    #[uniffi::constructor]
-    pub fn new() -> Self {
-        Self(iota_crypto::passkey::PasskeyVerifier::new())
-    }
-
-    pub fn verify(&self, message: &[u8], authenticator: &PasskeyAuthenticator) -> Result<()> {
-        Ok(self.0.verify(message, &authenticator.0)?)
-    }
-}
